Reorganising the CallerComponent code in multiple files
