+++ Grasshopper_UI/Components/oM/CreateCustom.cs
@@ -45,7 +45,7 @@ namespace BH.UI.Grasshopper.Components
 
         public CreateCustomComponent() : base()
         {
-            this.Params.ParameterChanged += OnGrasshopperUpdates;
+            this.Params.ParameterChanged += OnGHParamChanged;
         }
